initial commit
This is the first commit of the asynchronous "expector" python library.

This library is designed for writing tests that work something like
"expect" scripts, except that they can deal with multiple separate
"streams", which are text input and output from:

 - programs
 - sockets
 - ssh sessions
 - serial connections

Add-ons for dealing with other devices, like DMMs, PSUs, and
other hardware devices are also possible.